Refactor indexer_base -> base
