+++ medusa/indexers/tmdb/tmdb.py
@@ -12,7 +12,7 @@ from datetime import datetime, timedelta
 from dateutil import parser
 
 from medusa.app import TMDB_API_KEY
-from medusa.indexers.indexer_base import (Actor, Actors, BaseIndexer)
+from medusa.indexers.base import (Actor, Actors, BaseIndexer)
 from medusa.indexers.indexer_exceptions import IndexerError, IndexerException, IndexerShowIncomplete, IndexerUnavailable
 from medusa.logger.adapters.style import BraceAdapter
