+++ medusa/indexers/tvdbv2/tvdbv2_api.py
@@ -8,7 +8,7 @@ from collections import OrderedDict
 from medusa import app
 from medusa.app import TVDB_API_KEY
 from medusa.helper.metadata import needs_metadata
-from medusa.indexers.indexer_base import (Actor, Actors, BaseIndexer)
+from medusa.indexers.base import (Actor, Actors, BaseIndexer)
 from medusa.indexers.indexer_exceptions import (IndexerAuthFailed, IndexerError, IndexerException,
                                                 IndexerShowIncomplete, IndexerShowNotFound,
                                                 IndexerShowNotFoundInLanguage, IndexerUnavailable)
